requirements.txt
facebook-sdk==0.4.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
requests==2.7.0
Werkzeug==0.10.4
